requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
numpy~=1.26.4
openai~=1.40.1
protobuf==5.27.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
redis==5.0.7
regex==2024.7.24
requests==2.32.3
setuptools==71.1.0
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4

RAGatouille~=0.0.8.post2
pandas~=2.2.2
scikit-learn~=1.5.1
PyMuPDF~=1.24.9
pysbd~=0.3.4
umap-learn==0.5.5
